wavego_agent/requirements.txt
# =============================================================================
# WaveGo Agent - Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
#
# Note: Some packages may require additional system libraries.
# On Raspberry Pi, you may need:
#   sudo apt-get install python3-pyaudio portaudio19-dev espeak
# =============================================================================

# -----------------------------------------------------------------------------
# Core Dependencies
# -----------------------------------------------------------------------------
pyyaml>=6.0               # Configuration file parsing
numpy>=1.21.0             # Numerical operations

# -----------------------------------------------------------------------------
# Serial Communication (ESP32)
# -----------------------------------------------------------------------------
pyserial>=3.5             # Serial port communication

# -----------------------------------------------------------------------------
# Computer Vision
# -----------------------------------------------------------------------------
# IMPORTANT: Use opencv-contrib-python (NOT opencv-python) to match original setup.py
# They conflict if both installed - contrib is a superset with extra modules
opencv-contrib-python>=4.5.0  # OpenCV with extra modules (ArUco, tracking, etc.)
# For Raspberry Pi headless: opencv-contrib-python-headless>=4.5.0
imutils>=0.5.4                # OpenCV convenience functions

# -----------------------------------------------------------------------------
# Original Project Dependencies (from setup.py)
# -----------------------------------------------------------------------------
zmq>=25.0.0               # ZeroMQ messaging (used by original project)
pybase64>=1.2.0           # Base64 encoding (used by original project)

# -----------------------------------------------------------------------------
# Speech Recognition (STT)
# -----------------------------------------------------------------------------
SpeechRecognition>=3.8.0  # Google Speech Recognition
PyAudio>=0.2.11           # Audio input/output
# Note: PyAudio requires portaudio. Install with:
#   sudo apt-get install python3-pyaudio portaudio19-dev (Linux)
#   brew install portaudio && pip install pyaudio (macOS)

# Optional: Local speech recognition
# vosk>=0.3.32             # Offline speech recognition (optional)
# openai-whisper>=20230314 # Local Whisper model (optional, requires GPU)

# -----------------------------------------------------------------------------
# Text-to-Speech (TTS)
# -----------------------------------------------------------------------------
pyttsx3>=2.90             # Offline TTS engine
# Note: pyttsx3 requires espeak on Linux:
#   sudo apt-get install espeak

# Optional: Online TTS
gTTS>=2.3.0               # Google TTS (requires internet)
pygame>=2.1.0             # Audio playback for gTTS

# -----------------------------------------------------------------------------
# LLM API
# -----------------------------------------------------------------------------
openai>=1.0.0             # OpenAI API client

# -----------------------------------------------------------------------------
# Web Server (Optional - for video streaming)
# -----------------------------------------------------------------------------
flask>=2.0.0              # Web framework
flask-cors>=3.0.0         # CORS support
websockets>=10.0          # WebSocket support

# -----------------------------------------------------------------------------
# System Monitoring
# -----------------------------------------------------------------------------
psutil>=5.8.0             # System monitoring

# -----------------------------------------------------------------------------
# Development Dependencies (Optional)
# -----------------------------------------------------------------------------
# pytest>=7.0.0           # Testing
# black>=22.0.0           # Code formatting
# flake8>=4.0.0           # Linting
